requirements.txt
fastmcp >= 2.10.5
requests >= 2.32.4
python-dotenv >= 1.1.1
splunk-sdk >= 2.1.0